Update DW1000.cpp
Bug in clearAllStatus()- previous implementation has writing a value of 0 to every bit in the status register, while per the user manual latched bits in sys_status are cleared by writing 1 to them. clearReceiveStatus() and clearTransmitStatus() follow this by using setBit() to write a 1 to specific bits in sys_status. clearAllStatus() should accordingly write a 1 to every bit in sys_status, and can do so by calling memset with 0xff as the value to write to each byte.

When I tested the previous version of clearAllStatus() it resulted in no change to the internal status register of the DW1000 (which can be read by calling readSystemEventStatusRegister()). On the other hand, calling clearAllStatus() with 0xff as the value to write resulted in a full cleared (zeroed out) status register for the DW1000.