fix crash when searchDistantDevice returns NULL
This can occur when the tag and the anchor are at the edge of the
connection range as describe in issue #270.
This bug fix is actually an update of #236, but with one extra check
missing and the error message only printed in debug mode.
The configuration have been tested and shows that:
- before the fix, the anchor was doing a hard fault due to the NULL pointer
- after the fix, the ranging is properly working